+++ wflow-py/UnitTests/Test_w3ra_BMI.py
@@ -23,7 +23,7 @@ class MyTest(unittest.TestCase):
 
     def testbmirunnetcdfw3ra(self):
         bmiobj = bmi.wflowbmi_csdms()
-        bmiobj.initialize_config('../../examples/openstreams_w3ra_usa/wflow_w3ra.ini',loglevel=logging.DEBUG)
+        bmiobj.initialize_config('../../examples/openstreams_w3ra_usa/wflow_w3ra.ini',loglevel=logging.ERROR)
         bmiobj.set_attribute_value('run:runlengthdetermination','intervals')
 
         stime= calendar.timegm(parser.parse("2014-05-13 00:00:00").utctimetuple())
@@ -31,7 +31,7 @@ class MyTest(unittest.TestCase):
         bmiobj.set_start_time(stime)
         bmiobj.set_end_time(etime)
         st = bmiobj.get_start_time()
-        print st
+        #print st
         ett = bmiobj.get_end_time()
         ts = bmiobj.get_time_step()
 
@@ -44,9 +44,9 @@ class MyTest(unittest.TestCase):
             bmiobj.set_value('PRECIP',avar)
             cnt = cnt + 1
             bmiobj.update_until(curtime + ts)
-            print (curtime + ts)/ts
+            #print (curtime + ts)/ts
             curtime = bmiobj.get_current_time()
-            print bmiobj.get_current_time() - lastcurtime
+            #print bmiobj.get_current_time() - lastcurtime
             lastcurtime = bmiobj.get_current_time()