+++ wflow-py/wflow/__init__.py
@@ -1,4 +1,6 @@
 __all__ = ["wflow_funcs","wflow_adapt","wflow_lib","pcrut","wf_DynamicFramework","stats"]
 __version__="1.0.master"
 __release__="1.0.master.1"
+__versionnr__="1.0.1"
+__build__="2016-11-03 15:22:18.144000"
 import osgeo.gdal as gdal