+++ wflow-py/wflow/rtc_wflow_bmi.py
@@ -1,0 +1,10 @@
+
+from wrapperExtended import BMIWrapperExtended
+
+
+
+class rtcbmi_csdms(BMIWrapperExtended):
+    
+    
+    def initialize_model(self):
+        self.initialize('..')