+++ wflow-py/wflow/wf_DynamicFramework.py
@@ -17,6 +17,7 @@ import datetime
 import ConfigParser
 
 from wflow.wf_netcdfio import *
+import wflow
 import pcrut
 import glob
 import traceback
@@ -32,6 +33,7 @@ import calendar
 
 from wflow import __version__
 from wflow import __release__
+from wflow import __build__
 
 def log_uncaught_exceptions(ex_cls, ex, tb):
     global logging
@@ -96,6 +98,7 @@ class runDateTimeInfo():
                 self.runStateTime = self.runStartTime - datetime.timedelta(seconds=self.timeStepSecs)
             else:
                 self.runStateTime = self.runStartTime
+
             self.outPutStartTime = self.runStateTime + datetime.timedelta(seconds=self.timeStepSecs)
         elif timestepsecs and runTimeSteps:
             self.timeStepSecs = timestepsecs
@@ -198,6 +201,56 @@ class wf_exchnageVariables():
                     return 1
         return 1
 
+
+class wf_online_stats():
+    def __init__(self):
+        """
+
+        :param invarname:
+        :param mode:
+        :param points:
+        :param filename:
+        """
+        self.count = {}
+        self.rangecount= {}
+        self.result = {}
+        self.mode ={}
+        self.points = {}
+        self.filename = {}
+        self.statvarname = {}
+
+    def addstat (self, name, mode='mean', points=30, filename=None):
+        """
+
+        :param name:
+        :param mode:
+        :param points:
+        :param filename:
+        :return:
+        """
+        self.statvarname[name] = name + '_' + mode + '_' + str(points)
+        self.mode[name] = mode
+        self.points[name] = points
+        self.count[name] = 0
+        self.rangecount[name] = 0
+        self.filename[name] = filename
+
+    def getstat(self,data,name):
+        """
+
+        :param data:
+        :param name:
+        :return:
+        """
+        if self.count[name] == 0:
+            self.result[name] = data
+        else:
+            if self.mode[name] =='mean':
+                self.result[name] = self.result[name] * (self.points[name] -1)/self.points[name] + data/self.points[name]
+
+        self.count[name] = self.count[name] + 1
+
+        return scalar(self.result[name])
 
 class wf_sumavg():
     def __init__(self, varname, mode='sum', filename=None):
@@ -384,6 +437,9 @@ class wf_DynamicFramework(frameworkBase.FrameworkBase):
         self._addMethodToClass(self.wf_updateparameters)
         self._addAttributeToClass("ParamType", self.ParamType)
         self._addAttributeToClass("timestepsecs", self.DT.timeStepSecs)
+        self._addAttributeToClass("__version__", __version__)
+        self._addAttributeToClass("__release__", __release__)
+        self._addAttributeToClass("__build__", __build__)
         self.skipfirsttimestep = 0
 
         if firstTimestep == 0:
@@ -820,6 +876,8 @@ class wf_DynamicFramework(frameworkBase.FrameworkBase):
 
         self.logger = self._userModel().logger
 
+        self.logger.info("Initialise framework version: " + __version__ + "(" + __release__ + ")")
+
         global logging
         logging = self.logger
 
@@ -999,6 +1057,19 @@ class wf_DynamicFramework(frameworkBase.FrameworkBase):
 
 
 
+
+        # Add the on-lien statistics
+        self.onlinestat = wf_online_stats()
+
+        rollingvars = configsection(self._userModel().config, "rollingmean")
+        for thisvar in rollingvars:
+            try:
+                thisvarnoself = thisvar.split('self.')[1]
+            except:
+                logging.error("Entry in ini invalid: " + thisvar)
+                raise ValueError
+            pts = int(self._userModel().config.get("rollingmean", thisvar))
+            self.onlinestat.addstat(thisvarnoself,points=pts)
 
         # Fill the summary (stat) list from the ini file
         self.statslst = []
@@ -1985,11 +2056,18 @@ class wf_DynamicFramework(frameworkBase.FrameworkBase):
                 self._decrementIndentLevel()
                 # Save state variables in memory
                 self.wf_QuickSuspend()
-                self.wf_savedynMaps()
-                self.wf_saveTimeSeries()
+
                 for a in range(0, len(self.statslst)):
                     data = getattr(self._userModel(), self.statslst[a].varname)
                     self.statslst[a].add_one(data)
+
+                for key in self.onlinestat.statvarname:
+                    stvar = self.onlinestat.getstat(getattr(self._userModel(),key),key)
+                    #stvar = self.onlinestat.getstat(cover(self.DT.currentTimeStep * 1.0), key)
+                    setattr(self._userModel(),self.onlinestat.statvarname[key],stvar)
+
+                self.wf_savedynMaps()
+                self.wf_saveTimeSeries()
 
             #self.currentdatetime = self.currentdatetime + dt.timedelta(seconds=self._userModel().timestepsecs)
 
@@ -2286,7 +2364,9 @@ class wf_DynamicFramework(frameworkBase.FrameworkBase):
 
             if self._userModel()._inDynamic():
                 if 'None' not in self.ncfile:
-                    retval, succ = self.NcInput.gettimestep(self._userModel().currentTimeStep(), self.logger, var=varname,shifttime=self.DT.startadjusted)
+                    retval, succ = self.NcInput.gettimestep(self._userModel().currentTimeStep(), self.logger,
+                                                            tsdatetime=self.DT.currentDateTime + datetime.timedelta(seconds=self.DT.timeStepSecs), var=varname,
+                                                            shifttime=self.DT.startadjusted)
                     if succ:
                         return retval
                     else: