+++ wflow-py/wflow/wf_netcdfio.py
@@ -62,7 +62,7 @@ def convertCoord(proj_src, proj_trg, x, y):
 
 
 def prepare_nc(trgFile, timeList, x, y, metadata, logger, EPSG="EPSG:4326", units=None,
-               calendar='gregorian', Format="NETCDF4", complevel=9, zlib=True, least_significant_digit=None):
+               calendar='gregorian', Format="NETCDF4", complevel=9, zlib=True, least_significant_digit=None,FillValue=1E31):
     """
     This function prepares a NetCDF file with given metadata, for a certain year, daily basis data
     The function assumes a gregorian calendar and a time unit 'Days since 1900-01-01 00:00:00'
@@ -91,7 +91,7 @@ def prepare_nc(trgFile, timeList, x, y, metadata, logger, EPSG="EPSG:4326", unit
     else:
         nc_trg.createDimension('time', 0)  # NrOfDays*8
 
-    DateHour = nc_trg.createVariable('time', 'f8', ('time',), fill_value=-9999., zlib=zlib, complevel=complevel)
+    DateHour = nc_trg.createVariable('time', 'f8', ('time',), fill_value=FillValue, zlib=zlib, complevel=complevel)
     DateHour.units = units
     DateHour.calendar = calendar
     DateHour.standard_name = 'time'
@@ -113,12 +113,12 @@ def prepare_nc(trgFile, timeList, x, y, metadata, logger, EPSG="EPSG:4326", unit
     if srs.IsProjected() == 0: # ONly lat lon needed
         nc_trg.createDimension('lat', len(y))
         nc_trg.createDimension('lon', len(x))
-        y_var = nc_trg.createVariable('lat', 'f4', ('lat',), fill_value=-9999., zlib=zlib, complevel=complevel)
+        y_var = nc_trg.createVariable('lat', 'f4', ('lat',), fill_value=FillValue, zlib=zlib, complevel=complevel)
         y_var.standard_name = 'latitude'
         y_var.long_name = 'latitude'
         y_var.units = 'degrees_north'
         y_var.axis = 'Y'
-        x_var = nc_trg.createVariable('lon', 'f4', ('lon',), fill_value=-9999., zlib=zlib, complevel=complevel)
+        x_var = nc_trg.createVariable('lon', 'f4', ('lon',), fill_value=FillValue, zlib=zlib, complevel=complevel)
         x_var.standard_name = 'longitude'
         x_var.long_name = 'longitude'
         x_var.units = 'degrees_east'
@@ -132,12 +132,12 @@ def prepare_nc(trgFile, timeList, x, y, metadata, logger, EPSG="EPSG:4326", unit
     else:  # Assume regular grid in m
         nc_trg.createDimension('y', len(y))
         nc_trg.createDimension('x', len(x))
-        y_var = nc_trg.createVariable('y', 'f4', ('y',), fill_value=-9999., zlib=zlib, complevel=complevel)
+        y_var = nc_trg.createVariable('y', 'f4', ('y',), fill_value=FillValue, zlib=zlib, complevel=complevel)
         y_var.standard_name = 'projection_y_coordinate'
         y_var.long_name = 'y-coordinate in Cartesian system'
         y_var.units = 'm'
         y_var.axis = 'Y'
-        x_var = nc_trg.createVariable('x', 'f4', ('x',), fill_value=-9999., zlib=zlib, complevel=complevel)
+        x_var = nc_trg.createVariable('x', 'f4', ('x',), fill_value=FillValue, zlib=zlib, complevel=complevel)
         x_var.standard_name = 'projection_x_coordinate'
         x_var.long_name = 'x-coordinate in Cartesian system'
         x_var.units = 'm'
@@ -449,6 +449,16 @@ class netcdfinput():
         self.fstep = 0
         self.lstep = self.fstep + self.maxsteps
 
+        self.datetime = self.dataset.variables['time'][:]
+        if hasattr(self.dataset.variables['time'],'units'):
+            self.timeunits=self.dataset.variables['time'].units
+        else:
+            self.timeunits ='Seconds since 1970-01-01 00:00:00'
+        if hasattr(self.dataset.variables['time'], 'calendar'):
+            self.calendar= self.dataset.variables['time'].calendar
+        else:
+            self.calendar ='gregorian'
+        self.datetimelist=netCDF4.num2date(self.datetime,self.timeunits, calendar=self.calendar)
 
         try:
             self.x = self.dataset.variables['x'][:]
@@ -478,9 +488,20 @@ class netcdfinput():
         x = _pcrut.pcr2numpy(_pcrut.xcoordinate(_pcrut.boolean(_pcrut.cover(1.0))), NaN)[0, :]
         y = _pcrut.pcr2numpy(_pcrut.ycoordinate(_pcrut.boolean(_pcrut.cover(1.0))), NaN)[:, 0]
 
-        (self.latidx,) = logical_and(self.y >= y.min(), self.y <= y.max()).nonzero()
-        (self.lonidx,) = logical_and(self.x >= x.min(), self.x <= x.max()).nonzero()
+        #Get average cell size
+        acc = diff(x).mean() * 0.25
+        if self.flip:
+            (self.latidx,) = logical_and(self.y[::-1] +acc >= y.min(), self.y[::-1] <= y.max() + acc).nonzero()
+            (self.lonidx,) = logical_and(self.x + acc >= x.min(), self.x <= x.max() + acc).nonzero()
+        else:
+            (self.latidx,) = logical_and(self.y +acc >= y.min(), self.y <= y.max() + acc).nonzero()
+            (self.lonidx,) = logical_and(self.x +acc >= x.min(), self.x <= x.max() + acc).nonzero()
 
+        if len(self.lonidx) != len(x):
+            logging.error("error in determining X coordinates in netcdf...")
+
+        if len(self.latidx) != len(y):
+            logging.error("error in determining X coordinates in netcdf...")
 
         for var in vars:
             try:
@@ -491,7 +512,7 @@ class netcdfinput():
 
 
 
-    def gettimestep(self, timestep, logging, var='P', shifttime=False):
+    def gettimestep(self, timestep, logging, tsdatetime=None, var='P', shifttime=False):
         """
         Gets a map for a single timestep. reads data in blocks assuming sequential access
 
@@ -499,6 +520,7 @@ class netcdfinput():
         :var logging: python logging object
         :var var: variable to get from the file
         :var shifttime: is True start at 1 in the NC file (instead of 0)
+        :var tsdatetime: Assumed date/time of this timestep
 
 
             window = data[dpos,latidx.min():latidx.max()+1,lonidx.min():lonidx.max()+1]
@@ -507,6 +529,10 @@ class netcdfinput():
             ncindex = timestep
         else:
             ncindex = timestep - 1
+
+        if tsdatetime != None:
+            if tsdatetime != self.datetimelist[ncindex]:
+                logging.warn("Date/time does not match. Wanted " + str(tsdatetime) + " got " + str(self.datetimelist[ncindex]))
 
         if self.alldat.has_key(var):
             if ncindex == self.lstep:  # Read new block of data in mem