+++ doc/release_notes.rst
@@ -4,6 +4,10 @@ Release notes
 
 2016.04
 -------
+.. note::
+
+    Several none-backwards compatible changes will be part of this release. Use 2016.03 for older models
+
 + update soil names in sbm to the sbm2 names:
   + FirstZoneKsatVer -> KsatVer
   + FirstZoneMinCapacity -> SoilMinThickness
@@ -11,6 +15,8 @@ Release notes
   + FirstZoneCapacity -> SoilWaterCapacity
   + FirstZoneFlux -> SatWaterFlux
   + FirstZoneDepth -> SatWaterDepth
++ [model]reinit moved to [run]reinit (same for fewsrun) in all models
++ added [rollingmean] section in framework
 
 
 2016.03